+++ winpr/libwinpr/utils/wlog/PacketMessage.h
@@ -60,8 +60,8 @@ struct _wPcap
 	FILE* fp;
 	char* name;
 	BOOL write;
-	int file_size;
-	int record_count;
+	SSIZE_T file_size;
+	size_t record_count;
 	wPcapHeader header;
 	wPcapRecord* head;
 	wPcapRecord* tail;
